Capture errors, pretty print results on test.
Pretty-printed results can be used with `diff` to see the effect
of any OndaEDF code changes.

On test, gitignored files with the subset of results with no
eeg, ekg, emg and eog also get pretty printed, for easy inspection.
Some of these are normal: there may not be signals of those types
in the input EDF signals.